prism/requirements.txt
requests==2.31.0
pg8000==1.30.5
python-dotenv==1.0.0